refactoring + create pages

+++ backend/controllers/todoController.js
@@ -2,7 +2,7 @@ const Todo = require("../models/todoModel");
 
 exports.getAllTodos = async (req, res, next) => {
   try {
-    const result = await Todo.find();
+    const result = await Todo.find({ userId: req.user.id });
     res.status(200).json(result);
   } catch (err) {
     next(err);
@@ -16,7 +16,6 @@ exports.getTodo = async (req, res, next) => {
     res.status(200).json(result);
   } catch (err) {
     next(err);
-    // res.status(400).json(err);
   }
 };
 
@@ -24,15 +23,26 @@ exports.createTodo = async (req, res, next) => {
   try {
     const body = req.body;
     const newtask = body["task"];
-    const newTodo = await Todo.create({ task: newtask, completed: false });
+    const newTodo = await Todo.create({
+      task: newtask,
+      completed: false,
+      userId: req.user.id,
+    });
     res.status(200).json(newTodo);
   } catch (err) {
     next(err);
   }
 };
 
-exports.updateAllTodos = async (req, res, next) => {
-  res.status(200).json("Update All Todos");
+exports.completeAllTodos = async (req, res, next) => {
+  try {
+    const filter = { userId: req.user.id };
+    const updateDoc = { completed: true };
+    const result = await Todo.updateMany(filter, updateDoc);
+    res.status(200).json(result);
+  } catch (err) {
+    next(err);
+  }
 };
 
 exports.updateTodo = async (req, res, next) => {
@@ -48,7 +58,8 @@ exports.updateTodo = async (req, res, next) => {
 
 exports.deleteAllTodos = async (req, res, next) => {
   try {
-    const result = await Todo.deleteMany();
+    const result = await Todo.deleteMany({ userId: req.user.id });
+    
     res.status(200).json(result);
   } catch (err) {
     next(err);
